stock_analysis_backend/requirements.txt
alembic==1.14.1
click==8.1.8
colorama==0.4.6
greenlet==3.1.1
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
